Add day-night cycle and villager sleep
